fixing a compilation error

+++ logg.go
@@ -77,7 +77,7 @@ func startLoggerActor() {
 
 				if logger.l != nil {
 					logger.l.Println(msg)
-					logger.written += len(msg)
+					logger.written += int64(len(msg))
 				}
 			}
